domain fixes

+++ CalculatorApi/CalculatorApi.Domain/Services/ITaxService.cs
@@ -2,5 +2,5 @@ namespace CalculatorApi.Domain.Services;
 
 public interface ITaxService
 {
-    Task<decimal?> GetTax();
+    Task<double?> GetTax();
 }
